prevent sd logging when in quiet mode (+1 squashed commits)
Squashed commits:

[a4a1cdd5] fixed type conversion

+++ koboldcpp.py
@@ -540,8 +540,8 @@ def sd_generate(genparams):
     biggest = max(width,height)
     if biggest > reslimit:
         scaler = biggest / reslimit
-        width = width / scaler
-        height = height / scaler
+        width = int(width / scaler)
+        height = int(height / scaler)
         width = width - (width%64)
         height = height - (height%64)
